+++ app/config/security.yml
@@ -4,7 +4,7 @@
 security:
     providers:
         fos_userbundle:
-            id: fos_user.user_provider.username
+            id: sylius_user_provider
     encoders:
         FOS\UserBundle\Model\UserInterface: sha512
     firewalls:
