+++ features/frontend/user_login.feature
@@ -6,10 +6,21 @@ Feature: Sign in to the store
 
     Background:
         Given there are following users:
-            | email       | password | enabled |
-            | bar@foo.com | foo      | yes     |
+            | email        | username | password | enabled |
+            | bar@foo.com  | bar      | foo      | yes     |
+            | bar2@foo.com | bar2     | foo      | no     |
 
     Scenario: Log in with username and password
+        Given I am on the store homepage
+          And I follow "Login"
+         When I fill in the following:
+            | Email    | bar |
+            | Password | foo |
+          And I press "Login"
+         Then I should be on the store homepage
+          And I should see "Logout"
+
+    Scenario: Log in with email and password
         Given I am on the store homepage
           And I follow "Login"
          When I fill in the following:
@@ -19,7 +30,17 @@ Feature: Sign in to the store
          Then I should be on the store homepage
           And I should see "Logout"
 
-    Scenario: Log in with bad credentials
+    Scenario: Log in with username and bad password
+        Given I am on the store homepage
+          And I follow "Login"
+         When I fill in the following:
+            | Email    | bar |
+            | Password | bar         |
+          And I press "Login"
+         Then I should be on login page
+          And I should see "Bad credentials"
+
+    Scenario: Log in with email and bad password
         Given I am on the store homepage
           And I follow "Login"
          When I fill in the following:
@@ -36,7 +57,7 @@ Feature: Sign in to the store
          Then I should be on login page
           And I should see "Bad credentials"
 
-    Scenario: Trying to login as non existing user
+    Scenario: Trying to login as non existing username
         Given I am on the store homepage
           And I follow "Login"
          When I fill in the following:
@@ -45,3 +66,33 @@ Feature: Sign in to the store
           And I press "Login"
          Then I should be on login page
           And I should see "Bad credentials"
+
+    Scenario: Trying to login as non existing email
+        Given I am on the store homepage
+          And I follow "Login"
+         When I fill in the following:
+            | Email    | foo@bar.com |
+            | Password | bar         |
+          And I press "Login"
+         Then I should be on login page
+          And I should see "Bad credentials"
+
+    Scenario: Trying to login with disabled username
+        Given I am on the store homepage
+          And I follow "Login"
+         When I fill in the following:
+            | Email    | bar2 |
+            | Password | foo  |
+          And I press "Login"
+         Then I should be on login page
+          And I should see "User account is disabled"
+
+    Scenario: Trying to login with disabled email
+        Given I am on the store homepage
+          And I follow "Login"
+         When I fill in the following:
+            | Email    | bar2@foo.com |
+            | Password | foo          |
+          And I press "Login"
+         Then I should be on login page
+          And I should see "User account is disabled"
