+++ src/Sylius/Component/Core/Model/User.php
@@ -327,26 +327,4 @@ class User extends BaseUser implements UserInterface
 
         return $this;
     }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setEmail($email)
-    {
-        parent::setEmail($email);
-        $this->setUsername($email);
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setEmailCanonical($emailCanonical)
-    {
-        parent::setEmailCanonical($emailCanonical);
-        $this->setUsernameCanonical($emailCanonical);
-
-        return $this;
-    }
 }
