+++ src/Sylius/Component/Core/Model/UserWithoutUsername.php
@@ -1,0 +1,41 @@
+<?php
+/*
+* This file is part of the Sylius package.
+*
+* (c) Paweł Jędrzejewski
+*
+* For the full copyright and license information, please view the LICENSE
+* file that was distributed with this source code.
+*/
+
+namespace Sylius\Component\Core\Model;
+
+/**
+ * UserWithoutUsername model.
+ *
+ * @author Paweł Jędrzejewski <pjedrzejewski@diweb.pl>
+ */
+class UserWithoutUsername extends User
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function setEmail($email)
+    {
+        parent::setEmail($email);
+        $this->setUsername($email);
+
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setEmailCanonical($emailCanonical)
+    {
+        parent::setEmailCanonical($emailCanonical);
+        $this->setUsernameCanonical($emailCanonical);
+
+        return $this;
+    }
+}
